Separate module for subscription management
